# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIS-CUSTOMERS",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/communes": {
            "get": {
                "tags": [
                    "Communes"
                ],
                "summary": "Retrieve all communes",
                "description": "Get a list of all communes.",
                "operationId": "86e992d6ebe9077589081d2231fef66c",
                "responses": {
                    "200": {
                        "description": "A list of communes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Communes"
                ],
                "summary": "Create a new commune",
                "description": "Create a new commune with id_reg, description and status.",
                "operationId": "b61f1f5e46b8d31f100e6b61461ef7f2",
                "parameters": [
                    {
                        "name": "id_reg",
                        "in": "query",
                        "description": "Region ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Commune description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the commune",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A",
                                "I",
                                "trash"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_reg",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "id_reg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 90,
                                        "example": "New Commune Description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I",
                                            "trash"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commune created successfully"
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/communes/{id}": {
            "put": {
                "tags": [
                    "Communes"
                ],
                "summary": "Update data commune",
                "description": "Update an existing commune's id_reg, description and status.",
                "operationId": "2b6023411f0d9251e40f344b5add0053",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commune ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_reg",
                        "in": "query",
                        "description": "Region ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Commune description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the commune",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A",
                                "I",
                                "trash"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_reg",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "id_reg": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 90,
                                        "example": "Updated Commune Description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I",
                                            "trash"
                                        ],
                                        "example": "I"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commune updated successfully"
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commune not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commune not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Communes"
                ],
                "summary": "Delete commune",
                "description": "Delete an existing commune by ID.",
                "operationId": "2c9c8d3f22df0f2f2a5cacd680f06c15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commune ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commune deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commune deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commune not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commune not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}